backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
yfinance>=1.2.0
beautifulsoup4==4.12.2
earningscall==1.4.0
youtube-transcript-api>=1.2.4
yt-dlp>=2024.0
python-multipart==0.0.6
